# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "cy-performance API",
    "description": "Cycling analysis + training plan backend (read models, what-ifs, jobs).",
    "version": "0.0.1"
  },
  "paths": {
...
    "/v1/activities": {
      "get": {
        "tags": [
          "activities"
        ],
        "summary": "List Activities",
        "description": "Newest first; filter by date range, sport, rides only, ride class.",
        "operationId": "list_activities_v1_activities_get",
        "parameters": [
          {
            "name": "start",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start"
            }
          },
          {
            "name": "end",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End"
            }
          },
          {
            "name": "sport",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sport"
            }
          },
          {
            "name": "rides_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Rides Only"
            }
          },
          {
            "name": "classification",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Classification"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivityPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/activities/{activity_id}": {
      "get": {
        "tags": [
          "activities"
        ],
        "summary": "Get Activity",
        "description": "Stored per-ride analysis with its Explanation.",
        "operationId": "get_activity_v1_activities__activity_id__get",
        "parameters": [
          {
            "name": "activity_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Activity Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivityDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/activities/{activity_id}/streams": {
      "get": {
        "tags": [
          "activities"
        ],
        "summary": "Get Streams",
        "description": "Chart-ready streams (bucket-averaged; ≤ 5 000 points by default).",
        "operationId": "get_streams_v1_activities__activity_id__streams_get",
        "parameters": [
          {
            "name": "activity_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Activity Id"
            }
          },
          {
            "name": "resolution_s",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 600,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Resolution S"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StreamSeries"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/activities/{activity_id}/ride-log": {
      "get": {
        "tags": [
          "activities"
        ],
        "summary": "Ride Log",
        "description": "WORKOUT + RIDE.LOG text for a Strava description (deterministic; no poem).",
        "operationId": "ride_log_v1_activities__activity_id__ride_log_get",
        "parameters": [
          {
            "name": "activity_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Activity Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RideLogOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "activities"
        ],
        "summary": "Save Ride Log",
        "description": "Store the edited RIDE.LOG (e.g. with the acrostic poem); \"\" restores the generated one.",
        "operationId": "save_ride_log_v1_activities__activity_id__ride_log_post",
        "parameters": [
          {
            "name": "activity_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Activity Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RideLogSave"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RideLogOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/activities/{activity_id}/strava-description": {
      "post": {
        "tags": [
          "activities"
        ],
        "summary": "Strava Description",
        "description": "Preview (``confirm=false``) or write the ride's RIDE.LOG to Strava (flag-gated).",
        "operationId": "strava_description_v1_activities__activity_id__strava_description_post",
        "parameters": [
          {
            "name": "activity_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Activity Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StravaDescriptionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StravaDescriptionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/activities/{activity_id}/feedback": {
      "get": {
        "tags": [
          "activities"
        ],
        "summary": "Get Feedback",
        "description": "Post-ride RPE / feel: the web answer, else intervals.icu's.",
        "operationId": "get_feedback_v1_activities__activity_id__feedback_get",
        "parameters": [
          {
            "name": "activity_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Activity Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RideFeedbackOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "activities"
        ],
        "summary": "Save Feedback",
        "description": "Store post-ride RPE / feel; readiness is recomputed for the next day right away.",
        "operationId": "save_feedback_v1_activities__activity_id__feedback_post",
        "parameters": [
          {
            "name": "activity_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Activity Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RideFeedbackIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RideFeedbackOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/autopilot": {
      "post": {
        "tags": [
          "autopilot"
        ],
        "summary": "Run",
        "description": "Queue this profile's morning run; poll ``GET /v1/jobs/{id}`` for the report.",
        "operationId": "run_v1_autopilot_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AutopilotRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/autopilot/runs": {
      "get": {
        "tags": [
          "autopilot"
        ],
        "summary": "Runs",
        "description": "Saved run reports (scheduled and manual), newest first.",
        "operationId": "runs_v1_autopilot_runs_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AutopilotRunOut"
                  },
                  "title": "Response Runs V1 Autopilot Runs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ActivityDetail": {
        "properties": {
          "summary": {
            "$ref": "#/components/schemas/ActivitySummary"
          },
          "distance_m": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Distance M"
          },
          "elev_gain_m": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Elev Gain M"
          },
          "avg_w": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg W"
          },
          "avg_hr": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg Hr"
          },
          "max_hr": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Hr"
          },
          "vi": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vi"
          },
          "hr_lag_s": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hr Lag S"
          },
          "power_curve": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Power Curve"
          },
          "time_in_zone_power": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "number"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time In Zone Power"
          },
          "time_in_zone_hr": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "number"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time In Zone Hr"
          },
          "climbs": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Climbs"
          },
          "efforts": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Efforts"
          },
          "pacing": {
            "additionalProperties": true,
            "type": "object",
            "title": "Pacing"
          },
          "durability": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Durability"
          },
          "next_recommendation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Recommendation"
          },
          "explanation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Explanation"
              },
              {
                "type": "null"
              }
            ]
          },
          "has_streams": {
            "type": "boolean",
            "title": "Has Streams",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "summary"
        ],
        "title": "ActivityDetail",
        "description": "``GET /v1/activities/{id}``: the stored per-ride analysis."
      },
      "ActivityPage": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ActivitySummary"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "total",
          "offset",
          "limit",
          "items"
        ],
        "title": "ActivityPage",
        "description": "``GET /v1/activities``."
      },
...
      "AutopilotRequest": {
        "properties": {
          "write": {
            "type": "boolean",
            "title": "Write",
            "default": false
          },
          "confirm": {
            "type": "boolean",
            "title": "Confirm",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "AutopilotRequest",
        "description": "``POST /v1/autopilot``: run the morning now. ``write`` needs ``confirm``."
      },
      "AutopilotRunOut": {
        "properties": {
          "profile": {
            "type": "string",
            "title": "Profile"
          },
          "day": {
            "type": "string",
            "format": "date",
            "title": "Day"
          },
          "planner_mode": {
            "type": "string",
            "title": "Planner Mode"
          },
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "At"
          },
          "stages": {
            "items": {
              "$ref": "#/components/schemas/AutopilotStage"
            },
            "type": "array",
            "title": "Stages"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "profile",
          "day",
          "planner_mode",
          "ok"
        ],
        "title": "AutopilotRunOut",
        "description": "``GET /v1/autopilot/runs``: a saved run report."
      },
      "AutopilotStage": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "detail": {
            "type": "string",
            "title": "Detail",
            "default": ""
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "status"
        ],
        "title": "AutopilotStage",
        "description": "One stage of an autopilot run."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RideFeedbackIn": {
        "properties": {
          "rpe": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rpe"
          },
          "feel": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Feel"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "RideFeedbackIn",
        "description": "``POST /v1/activities/{id}/feedback``: post-ride RPE 1–10 and feel 1 (strong) … 5 (weak)."
      },
      "RideFeedbackOut": {
        "properties": {
          "activity_id": {
            "type": "integer",
            "title": "Activity Id"
          },
          "rpe": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rpe"
          },
          "feel": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Feel"
          },
          "source": {
            "type": "string",
            "title": "Source"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "activity_id",
          "source"
        ],
        "title": "RideFeedbackOut",
        "description": "Answers for a ride and where they came from (``web`` / ``intervals.icu`` / ``none``)."
      },
      "RideLogOut": {
        "properties": {
          "activity_id": {
            "type": "integer",
            "title": "Activity Id"
          },
          "workout": {
            "type": "string",
            "title": "Workout"
          },
          "ride_log": {
            "type": "string",
            "title": "Ride Log"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "source": {
            "type": "string",
            "title": "Source",
            "default": "generated"
          },
          "generated": {
            "type": "string",
            "title": "Generated",
            "default": ""
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "activity_id",
          "workout",
          "ride_log",
          "text"
        ],
        "title": "RideLogOut",
        "description": "``GET /v1/activities/{id}/ride-log``: the saved text (with poem) or the generated one."
      },
      "RideLogSave": {
        "properties": {
          "text": {
            "type": "string",
            "maxLength": 10000,
            "title": "Text"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "text"
        ],
        "title": "RideLogSave",
        "description": "``POST /v1/activities/{id}/ride-log``: store an edited text (\"\" = back to generated)."
      },
...
      "StravaDescriptionOut": {
        "properties": {
          "strava_id": {
            "type": "integer",
            "title": "Strava Id"
          },
          "before": {
            "type": "string",
            "title": "Before"
          },
          "after": {
            "type": "string",
            "title": "After"
          },
          "written": {
            "type": "boolean",
            "title": "Written"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "strava_id",
          "before",
          "after",
          "written"
        ],
        "title": "StravaDescriptionOut",
        "description": "Result of a preview or a write."
      },
      "StravaDescriptionRequest": {
        "properties": {
          "text": {
            "type": "string",
            "maxLength": 10000,
            "title": "Text"
          },
          "confirm": {
            "type": "boolean",
            "title": "Confirm",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "text"
        ],
        "title": "StravaDescriptionRequest",
        "description": "``POST /v1/activities/{id}/strava-description``: preview, or write with ``confirm``."
      },
      "StreamSeries": {
        "properties": {
          "activity_id": {
            "type": "integer",
            "title": "Activity Id"
          },
          "resolution_s": {
            "type": "integer",
            "title": "Resolution S"
          },
          "n": {
            "type": "integer",
            "title": "N"
          },
          "columns": {
            "additionalProperties": {
              "items": {
                "anyOf": [
                  {
                    "type": "number"
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "type": "array"
            },
            "type": "object",
            "title": "Columns"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "activity_id",
          "resolution_s",
          "n",
          "columns"
        ],
        "title": "StreamSeries",
        "description": "``GET /v1/activities/{id}/streams``: column arrays, downsampled for charts."
      },
...
    }
  }
}